refactor(iocuak-core): rename task module to createInstanceTask
